MGMT-20908: Add support for the OAuth authorization code flow
Currently the server requires the offine token in the `OFFLINE_TOKEN`
environment variable, or in a request header. But use of offline tokens
is deprected, and will be removed in the future. To avoid using them
this patch adds an optional mechanism to use the OAuth authorization
code flow, intended for use when the server is executed locally by the
user. When the server starts it checks if the
`USE_AUTHORIZATION_CODE_FLOW` environment variable is set to `true`. If
it is then the authorization URL will be opened with the local browser,
so that the user can provide the credentials. Then the authorization
server will send the authorization code to the `/oauth/callback`
endpoint, and the MCP sever will exchange that code for the refresh and
access tokens. The refresh token is then used as it was the offline
token.

Related: https://issues.redhat.com/browse/MGMT-20908
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>