Implement OAuth endpoints
This patch implements enough OAuth endpoints to make the server usable
from MCP clients like VScode without having to manually provide a token.

When the MCP client receives a 401 code from the server it will assume
that it uses OAuth, and it will try to fetch the OAuth metadata from
`/.well-known/oauth-authorization-server`. As we don't want to filly
implement that, we implement it calling the equivalent endpoint of
_sso.redhat.com_. We then modify a couple of things in the content, and
return it to the client.

One of the things that we modify is the dynamic client registration
endpoint. The MCP client will try to use the one in the metadata, but
_sso.redhat.com_ doesn't allow that. So we replace it with our own,
where we accept all requests and respond always with the same client
identifier.

With this the MCP client is able to complete the OAuth process, and
obtain an access token that will then use to call the MCP. The MCP will
then in turn use it to call the assisted installer REST API.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>