Return less on error
Before this, it was possible that internal implementation details in
error messages would be returned to the LLM when the MCP server
encountered an error. This isn't what we want because users shouldn't
get messages from the LLM about the application.

This continues to log issues, while returning just a minimal response
message to the agent on error.

https://issues.redhat.com/browse/MGMT-21113