Add operator names to tool description
Before this asking the bot to add an operator by name would cause it to
attempt to call `add_operator_bundle_to_cluster` with an invalid name.
Listing the available operators would solve the issue, but since we only
allow two operator bundles it's easier to just add the valid names to
the description.

Listing the bundles will still give more information about them like
which individual operators are included in the bundle if the user needs
that info.

https://issues.redhat.com/browse/MGMT-21356