Fix logging and requests sync calls
The blockbuster tool revealed that both logging and the use of the
requests  library for getting the token and pull secret were blocking
the async nature of the mcp server.

This commit handles the logging by splitting the log into a producer and
consumer through a queue and using async.to_thread for the requests
calls.