Static Networking Config
Add tools for CRUD operations on a host's static network config.

Also add a tool to validate the static network nmstate YAML, and a tool
to generate an initial nmstate YAML for the most common configurations
used.

The API handling is a bit awkward given the MAC to IP mapping but
all of that complexity is hidden from the LLM.