Add a signature container crashes
This signature detects when containers are crashing during bootstrap on
any of the bootstrap or control plane nodes and puts the last 20 lines
of those containers into the output.

Only crashes in the last 30 minutes of the install are counted and
crashes are detected by scanning kubelet.log for lines indicating a
CrashLoopBackOff condidion.