Fix auth test contamination
After the move a change in import order caused this test to start
failing. This is because the settings are a global concept and end up
shared across tests. This should be fixed more generally, but for now
just isolate the import of auth in the auth tests to address the
test-specific problem.